+++ backend/internal/model/role.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Role struct {
-	ID        uint8 `gorm:"primaryKey"`
+	ID        uint16 `gorm:"primaryKey"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 	Name      string `gorm:"type:varchar(150);unique;check:length(trim(name)) >= 3"`
